Moved returnMaxUploadSize to the functions.php
