+++ inc/functions.php
@@ -1,4 +1,4 @@
-<?php
+﻿<?php
 /********************************************************************************
 *                                                                               *
 *   Copyright 2012 Nicolas CARPi (nicolas.carpi@gmail.com)                      *
@@ -43,6 +43,29 @@ function kdate()
         $day = '0' . $day;
     }
     return $today['year'] . $month . $day;
+}
+
+/**
+ * Converts the php.ini upload size setting to a numeric value in MB
+ * Returns 2 if no value is found (utilizing the default setting that was in there previously)
+*/
+function returnMaxUploadSize() {    
+    $val = trim($ini_get('upload_max_filesize'));
+    
+    if (!isset($val)) {
+      return 2;
+    }
+    
+    $last = strtolower($val[strlen($val)-1]);
+    
+    switch($last) {
+        case 'g':
+            $val *= 1000;
+        case 'k':
+            $val /= 1024;
+    }
+
+    return $val;
 }
 
 /**
